Add files via upload
this is a university project, the aim of the program is to create a calculator that calculate the sale. In the first text field is used to insert the original price the second text field is used to input the sale and the last text field is use to to display the final price.